Feat: EmojiDropDown 컴포넌트 분리 및 구현
